+++ Changelog.md
@@ -30,6 +30,7 @@
 - Adds basic support for Indicator status line and their VT sequences `DECSASD` and `DECSSDT`, and `DECRQSS` has been adapted (#687).
 - Adds configuration option `profiles.*.status_line.display` to be either `none` or `indicator` to reflect the initial state of the status line (more customizability of the Indicator status-line will come in future releases).
 - Adds new action `ToggleInputProtection` to protect terminal application against accidental input (#697).
+- Adds configuration options `logging.enabled` as well as `logging.file`.
 - Changes CLI syntax for `contour parser-table` to `contour generate parser-table`.
 
 ### 0.3.1 (2022-05-01)